requirements.txt
# GP Triage POC - Dependencies
# Compatible with Python 3.12+

# Core framework
streamlit==1.29.0

# LangChain ecosystem - 0.2.x series (pydantic v2 native, Python 3.12 compatible)
langchain>=0.2.0,<0.3
langchain-openai>=0.1.8,<0.2
langchain-community>=0.2.0,<0.3
langchain-core>=0.2.0,<0.3
langchain-chroma>=0.1.0,<0.2

# OpenAI
openai>=1.14.0,<2.0

# Vector database
chromadb==0.4.22

# Utilities
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.4

# Pydantic v2
pydantic>=2.0.0

# HTTP client (explicit injection to avoid httpx proxies= crash)
httpx>=0.27.0,<1.0

# Development/Testing
pytest==7.4.3
black==23.12.1
